Add third_party and update CMakeLists.txt

+++ scripts/build_light_bolt.sh
@@ -7,6 +7,7 @@ BOLT_ROOT=${script_dir}/..
 current_dir=${PWD}
 
 build_dir=$1
+use_mali=$2
 
 srcs=""
 searchFiles() {
@@ -33,13 +34,15 @@ searchFiles ${build_dir}
 srcs="${srcs} ${build_dir}/model-tools/src/CMakeFiles/model-tools.dir/model_serialize_deserialize.cpp.o \
 ${build_dir}/model-tools/src/CMakeFiles/model-tools.dir/model_tools.cpp.o"
 
-if [ -f "${BOLT_ROOT}/third_party/llvm/opencl/lib64/libOpenCL.so" ];then
+if [ -f "${BOLT_ROOT}/third_party/llvm/opencl/lib64/libOpenCL.so" ] && [ $use_mali == "ON" ];
+then
     cp ${BOLT_ROOT}/third_party/llvm/opencl/lib64/libOpenCL.so ${build_dir}
     aarch64-linux-android-strip ${build_dir}/libOpenCL.so || exit 1
     #aarch64-linux-android-readelf -dW ${build_dir}/libOpenCL.so
 fi
 
-if [ -f "${BOLT_ROOT}/gcl/tools/kernel_lib_compile/lib/libkernelbin.so" ];then
+if [ -f "${BOLT_ROOT}/gcl/tools/kernel_lib_compile/lib/libkernelbin.so" ] && [ $use_mali == "ON" ];
+then
     aarch64-linux-android-strip ${BOLT_ROOT}/gcl/tools/kernel_lib_compile/lib/libkernelbin.so || exit 1
     #cp ${BOLT_ROOT}/gcl/tools/kernel_lib_compile/lib/libkernelbin.so ${build_dir}
     aarch64-linux-android21-clang++ -shared -o ${build_dir}/libBoltModel.so ${srcs} \
